requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.0
pytest
